+ Fixed cast bug

+++ src/main/java/fi/uef/envi/wavellite/entity/derivation/base/AbstractComponentPropertyValue.java
@@ -45,7 +45,7 @@ public abstract class AbstractComponentPropertyValue extends AbstractEntity
 		if (value == null)
 			throw new NullPointerException("[value = null]");
 
-		this.value = (Double) value;
+		this.value = value;
 	}
 
 	@Override
